Adding some more definition around the Expense model

+++ NETS/projects/commons-user/trunk/commons-user-java-lib/src/main/java/com/neosavvy/user/dao/base/BaseCompanyDAO.java
@@ -34,7 +34,7 @@ public interface BaseCompanyDAO<T extends AbstractCompany> {
 
     public T updateCompany(T company);
 
-	public T findCompanyById(int id);
+	public T findCompanyById(long id);
 
 	public List<T> findCompanies(T company);
