+++ NETS/projects/commons-user/trunk/commons-user-java-lib/src/main/java/com/neosavvy/user/dao/companyManagement/RoleDAO.java
@@ -16,7 +16,7 @@ public interface RoleDAO {
 
 	public void saveRole(RoleDTO user);
 
-	public RoleDTO findRoleById(int id);
+	public RoleDTO findRoleById(long id);
 
 	public List<RoleDTO> findRoles(RoleDTO role);
